+++ src/pigrun/SlidingTopObstacleUpWard.java
@@ -6,7 +6,7 @@ public class SlidingTopObstacleUpWard extends TopObstacle {
 
 	
 
-	public SlidingTopObstacleUpWard(float x, float y1, float vx) throws SlickException {
+	public SlidingTopObstacleUpWard(float x, float y1, double vx) throws SlickException {
 		super(x, y1, vx);
 	
 	}
